Organize unittests better
Cherry-picked from https://github.com/GooborgStudios/mdn-bcd-collector/commit/60fe9c95bb537607892dcfcea9b5d060cdf61ea8.
