+++ unittest/unit/find-missing-features.ts
@@ -12,8 +12,10 @@ import fs from 'fs-extra';
 
 import {traverseFeatures, getMissing} from '../../find-missing-features.js';
 
-import bcd from './bcd.test.js';
-const tests = await fs.readJson(new URL('./tests.test.json', import.meta.url));
+import bcd from '../sample/bcd.test.js';
+const tests = await fs.readJson(
+  new URL('../sample/tests.test.json', import.meta.url)
+);
 
 describe('find-missing-features', () => {
   describe('traverseFeatures', () => {
